+++ packages/tinacms/src/toolkit/fields/plugins/mdx-field-plugin/plate/components/plate-ui/icons.tsx
@@ -78,6 +78,26 @@ const RawMarkdown = () => {
     </svg>
   )
 }
+const MermaidIcon = () => (
+  <svg
+    width="100%"
+    height="100%"
+    viewBox="0 0 491 491"
+    version="1.1"
+    xmlns="http://www.w3.org/2000/svg"
+    fillRule="evenodd"
+    clipRule="evenodd"
+    strokeLinejoin="round"
+    strokeMiterlimit={2}
+  >
+    <path d="M490.16,84.61C490.16,37.912 452.248,0 405.55,0L84.61,0C37.912,0 0,37.912 0,84.61L0,405.55C0,452.248 37.912,490.16 84.61,490.16L405.55,490.16C452.248,490.16 490.16,452.248 490.16,405.55L490.16,84.61Z" />
+    <path
+      d="M407.48,111.18C335.587,108.103 269.573,152.338 245.08,220C220.587,152.338 154.573,108.103 82.68,111.18C80.285,168.229 107.577,222.632 154.74,254.82C178.908,271.419 193.35,298.951 193.27,328.27L193.27,379.13L296.9,379.13L296.9,328.27C296.816,298.953 311.255,271.42 335.42,254.82C382.596,222.644 409.892,168.233 407.48,111.18Z"
+      fill="white"
+      fillRule="nonzero"
+    />
+  </svg>
+)
 
 const borderAll = (props: LucideProps) => (
   <svg
@@ -314,6 +334,7 @@ export const Icons = {
   clear: X,
   close: X,
   // code: Code2,
+  paint: PaintBucket,
   codeblock: FileCode,
   color: Baseline,
   column: RectangleVertical,
@@ -339,6 +360,7 @@ export const Icons = {
   lineHeight: WrapText,
   // link: Link2,
   minus: Minus,
+  mermaid: MermaidIcon,
   more: MoreHorizontal,
   // ol: ListOrdered,
   outdent: Outdent,