+++ playwright/tina-playwright/playwright.config.ts
@@ -5,7 +5,10 @@ import { defineConfig, devices } from "@playwright/test";
  */
 export default defineConfig({
   timeout: 10000,
-
+  reporter: [
+    ["list", { printSteps: true }],
+    ["json", { outputFile: "playwright-test-results.json" }],
+  ],
   testDir: "./tests",
   /* Run tests in files in parallel */
   fullyParallel: true,
@@ -16,7 +19,6 @@ export default defineConfig({
   /* Opt out of parallel tests on CI. */
   workers: process.env.CI ? 1 : undefined,
   /* Reporter to use. See https://playwright.dev/docs/test-reporters */
-  reporter: "html",
   /* Shared settings for all the projects below. See https://playwright.dev/docs/api/class-testoptions. */
   use: {
     trace: "on-first-retry",