+++ packages/tinacms-gitprovider-github/CHANGELOG.md
@@ -1,5 +1,19 @@
 # tinacms-gitprovider-github
 
+## 2.0.7
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tinacms/datalayer@1.3.7
+
+## 2.0.6
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tinacms/datalayer@1.3.6
+
 ## 2.0.5
 
 ### Patch Changes